requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
datapunt-keycloak-oidc==0.5.1
dj-rest-auth==2.2.7
Django==3.2.16
django-cleanup==6.0.0
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-maintenance-mode==0.18.0
django-mysql==3.11.1
django-qr-code==3.1.1
django-reorder-admin==0.3.1
django-resized==1.0.2
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
docopt==0.6.2
drf-tracking==1.5.0
et-xmlfile==1.1.0
fonttools==4.38.0
idna==3.4
image==1.5.33
josepy==1.13.0
kiwisolver==1.4.4
ldap3==2.9.1
mailjet-rest==1.3.4
matplotlib==3.6.3
mozilla-django-oidc==1.2.4
ndg-httpsclient==0.5.1
num2words==0.5.12
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pusher==3.3.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.1.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-fsutil==0.9.3
pytz==2022.7.1
qrcode==7.3.1
requests==2.28.2
rjsmin==1.1.0
schedule==1.1.0
segno==1.5.2
six==1.16.0
sqlparse==0.4.3
suds-community==0.8.5
tesseract==0.1.3
textdistance==4.5.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
